// src/app/models/contact-details.ts
export interface contactDetails {
    _id: string
    first_name: string
    last_name: string
    phone_number: string
    email: string
    address: string
    city: string
    state: string
    zip_code: Number
    country: string
    summary: string
    linkedin_url: string
    website_url: string
}